Refactoring tests to separate folders
